Final Game Playable Demo
Thanks for taking over to course to allow us to finish even with limited time. I had a lot of fun and learned a lot in this class.

I had a bunch of ideas for this game that didn't quite make it to the final version, but I hope you're able to see the vision.